Define an interface for LocateFileOrFolder
