+++ src/Helper/FindApplicationRoot/FindApplicationRootWithConfigOverride.php
@@ -6,14 +6,14 @@ namespace Scoutapm\Helper\FindApplicationRoot;
 
 use Scoutapm\Config;
 use Scoutapm\Config\ConfigKey;
-use Scoutapm\Helper\LocateFileOrFolder;
+use Scoutapm\Helper\LocateFileOrFolder\LocateFileOrFolder;
 use Scoutapm\Helper\Superglobals\Superglobals;
 
 use function array_key_exists;
 use function is_string;
 
 /** @internal This is not covered by BC promise */
-class FindApplicationRootWithConfigOverride implements FindApplicationRoot
+final class FindApplicationRootWithConfigOverride implements FindApplicationRoot
 {
     /** @var LocateFileOrFolder */
     private $locateFileOrFolder;
